Added movies & media sections are auto discovered

+++ custom_components/sensor/plex_recent_media.py
@@ -11,7 +11,7 @@ from homeassistant.components.sensor import PLATFORM_SCHEMA
 from homeassistant.const import CONF_HOST, CONF_PORT, CONF_MONITORED_CONDITIONS
 from homeassistant.helpers.entity import Entity
 
-__version__ = '0.0.2'
+__version__ = '0.0.3'
 
 _LOGGER = logging.getLogger(__name__)
 
@@ -29,7 +29,7 @@ SENSOR_TYPES = {
 
 ENDPOINTS = {
     'media':
-        'http://{0}:{1}/library/sections/1/recentlyAdded?X-Plex-Token={2}&X-Plex-Container-Start=0&X-Plex-Container-Size={3}'
+        'http://{0}:{1}/library/sections/{2}/recentlyAdded?X-Plex-Token={3}&X-Plex-Container-Start=0&X-Plex-Container-Size={4}'
 }
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend({
@@ -64,6 +64,7 @@ class Plex_RecentSensor(Entity):
         self._name = SENSOR_TYPES[self.type][0]
         self.attribNum = 0
         self.now = str(get_date(self._tz))
+        self.sections = []
 
     @property
     def name(self):
@@ -77,10 +78,11 @@ class Plex_RecentSensor(Entity):
 
     @property
     def device_state_attributes(self):
-        """Return the state attributes of the sensor."""
+        """Return JSON for the sensor."""
         data = []
         attributes = {}
         default = {}
+        self.attribNum = 0
         default['title_default'] = '$title'
         default['line1_default'] = '$episode'
         default['line2_default'] = '$release'
@@ -88,75 +90,84 @@ class Plex_RecentSensor(Entity):
         default['line4_default'] = '$number - $studio'
         default['icon'] = 'mdi:eye-off'
         data.append(default)
-        for show in self.data:
+        for show in self.data[:self.items]:
             pre = {}
-            self.attribNum += 1
-            if 'studio' in show:
+            directory = '../local/custom-lovelace/upcoming-media-card/images/plex/'
+            if show.get('type') == 'movie':
+                self.attribNum += 1
                 pre['title'] = show.get('title', '')
                 pre['episode'] = ''
-            else:
+            elif show.get('type') == 'episode':
+                self.attribNum += 1
                 pre['title'] = show.get('grandparentTitle', '')
                 pre['episode'] = show.get('title', '')
-            try:
-                if 'viewCount' in show: pre['flag'] = False
-                else: pre['flag'] = True
-            except: pre['flag'] = False
-            if show.get('rating', 0) > 0:
-                pre['rating'] = '\N{BLACK STAR}' + str(show.get('rating', ''))
-            else:
-                pre['rating'] = ''
-            pre['number'] = ('S{:02d}E{:02d}').format(show.get('parentIndex', ''), show.get('index', ''))
+                pre['number'] = ('S{:02d}E{:02d}').format(show.get('parentIndex', ''), show.get('index', ''))
+            else: continue
+            if 'viewCount' in show: pre['flag'] = False
+            else: pre['flag'] = True
+            if show.get('rating', 0) > 0: pre['rating'] = '\N{BLACK STAR}' + str(show['rating'])
+            else: pre['rating'] = ''
             pre['airdate'] = datetime.utcfromtimestamp(show.get('addedAt')).isoformat() + 'Z'
-            if math.floor(show.get('duration') / 60000) > 0:
-                pre['runtime'] = math.floor(show.get('duration') / 60000)
+            if show.get('duration', 0) > 0: pre['runtime'] = math.floor(show['duration'] / 60000)
+            try: pre['poster'] = '{0}p{1}.jpg'.format(directory,show['ratingKey'])
+            except: pre['poster'] = ''
+            try: pre['fanart'] = '{0}f{1}.jpg'.format(directory,show['ratingKey'])
+            except: pre['fanart'] = ''
             pre['release'] = '$day, $date $time'
-            try:
-                pre['poster'] = '../local/custom-lovelace/upcoming-media-card/images/plex/p' + show['ratingKey'] + '.jpg'
-            except:
-                pre['poster'] = ''
-            try:
-                pre['fanart'] = '../local/custom-lovelace/upcoming-media-card/images/plex/f' + show['ratingKey'] + '.jpg'
-            except:
-                pre['fanart'] = ''
             data.append(pre)
-
+        self._state = self.attribNum
         attributes['data'] = json.dumps(data)
         return attributes
 
     def update(self):
         """Update the data for the sensor."""
         try:
-            res = requests.get(
-                ENDPOINTS[self.type].format(
-                    self.host, self.port,
-                    self.token, self.items),
+            self.sections = []
+            getsec = requests.get('http://{0}:{1}/library/sections/all/?X-Plex-Token={2}'.format(
+                    self.host, self.port, self.token),
                 headers={ 'Accept': 'application/json' },
                 timeout=10)
+            for key in getsec.json()['MediaContainer']['Directory']:
+                self.sections.append(key['key'])
+
         except OSError:
             _LOGGER.warning("Host %s is not available", self.host)
             self._state = None
             return
         
-        if res.status_code == 200:
-            self.data = res.json()['MediaContainer']['Metadata']
-            self._state = self.attribNum
+        if getsec.status_code == 200:
+            self.data = []
+            for key in self.sections:
+                res = requests.get(
+                    ENDPOINTS[self.type].format(
+                        self.host, self.port, key,
+                        self.token, self.items),
+                    headers={ 'Accept': 'application/json' },
+                    timeout=10)
+                self.data += res.json()['MediaContainer']['Metadata']
 
         directory = 'www/custom-lovelace/upcoming-media-card/images/plex/'
         if not os.path.exists(directory): os.makedirs(directory)
 
         media_ids = []
-        for show in self.data: media_ids.append(show['ratingKey'])
+        self.data = sorted(self.data, key = lambda i: i['addedAt'], reverse=True)
+        for show in self.data[:self.items]:
+            media_ids.append(show['ratingKey'])
 
         """Get imgs for shown media, delete imgs for media no longer in list"""
         if not set(media_ids).issubset(os.listdir(directory)):
             for filename in os.listdir(directory):
-                if str(media_ids).find(filename[1:-4]) == -1: 
+                if filename.endswith('.jpg') and str(media_ids).find(filename[1:-4]) == -1: 
                     os.remove(directory + filename)
-            for show in self.data:
+
+            for show in self.data[:self.items]:
+                if show.get('type') == 'movie': poster = show.get('thumb')
+                elif show.get('type') == 'episode': poster = show.get('parentThumb')
+                else: continue
                 if not os.path.isfile(directory + 'p' + show['ratingKey'] + '.jpg'):
                     try:
                         imgurl = BytesIO(requests.get('http://{0}:{1}{2}?X-Plex-Token={3}'.format(
-                            self.host, self.port, show.get('parentThumb'), self.token)).content)
+                            self.host, self.port, poster, self.token)).content)
                         img = Image.open(imgurl)
                         img.resize((150, 225), Image.ANTIALIAS).save(
                             directory + 'p' + show['ratingKey'] + '.jpg', 'JPEG')
